Adding QuaMeter task.
Many changes related to creating distinct files.
If we want to call e.g. myrimatch twice, with identical parameters, we want it to run only once with a single output file.
The current setup would make MyriMatch produce two different files (distinct file takes care of that),
which will throw the rest of the pipeline out of whack.