Fixing a bug in quameter line colors
When instruments were shift-clicked out of order (first select 3rd, then select 1st), the selectedInstruments array contained out of order instrument names. That lead to mismatching colors in the plots.

The fix creates allInstruments array containing proper order of instruments. When a button is clicked, the instrument name is added to the selectedInstruments array, which is then subsequently sorted to match the allInstruments array order.